refactor: alterando estrutura de pacotes dos componentes de board
