+++ frontend/src/app/features/pages/game/components/game-session/components/board/components/square/square.scss
